+++ src/main/java/edu/odu/cs/cs350/AudioFile.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import edu.odu.cs.cs350.Enum.FileType;
 
 public class AudioFile extends WebsiteFile {
-	public AudioFile(int inSize, Path inPath) {
+	public AudioFile(long inSize, Path inPath) {
 		super(inSize, inPath, FileType.AUDIO);
 	}
