+++ src/main/java/edu/odu/cs/cs350/VideoFile.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import edu.odu.cs.cs350.Enum.FileType;
 
 public class VideoFile extends WebsiteFile {
-	public VideoFile(int inSize, Path inPath) {
+	public VideoFile(long inSize, Path inPath) {
 		super(inSize, inPath, FileType.VIDEO);
 	}
